test(config): move resolution tests into their own module
@orhun asked for the tests to live outside lib.rs. resolve_config_path
and its tests now sit in git-cliff/src/config_path.rs, matching how
args.rs keeps its tests alongside the code they cover.
